+++ _extensions/sverto/sverto-prerender.lua
@@ -1,62 +1,26 @@
+-- sverto-prerender.lua
+-- james goldie
+-- this pre-render script runs when sverto is installed on a quarto project. it:
+-- 1) looks in the doc meta for svelte paths
+-- 2) calls the svelte compiler with the svelte paths identified in (1)
+-- (injecting the bundled svelte files into the html is handled by sverto.lua)
 
--- return contents of named file
-function read_file(name)
-  local file = io.open(name, "r")
-  if file == nil then
-      return ""
+-- try to load a module from multiple paths
+-- modified version of https://stackoverflow.com/a/17878208/3246758
+local function try_require(path_table)
+  for _, v in pairs(path_table) do
+    ok, err = pcall(require, v)
+    if ok then
+      return err
+    end
   end
-  local contents = file:read("a")
-  file:close()
-  return contents
+  error("Could not find a module to load from selected paths")
 end
 
--- get a prefix for calling npm run based on windows or *nix/macos
--- using path separator: i don't know pandoc.system.os() values
-function get_cmd_prefix()
-  if pandoc.path.separator == "\\" then
-    return "cmd /c "
-  else
-    return ""
-  end
-end
-
--- file_exists: true if the file at `name` exists
--- from https://pandoc.org/lua-filters.html#building-images-with-tikz
-function file_exists(name)
-  local f = io.open(name, 'r')
-  if f ~= nil then
-    io.close(f)
-    return true
-  else
-    return false
-  end
-end
-
--- offset a relative `svelte_path` to a .qmd `input_path`, or an absolute
--- `svelte_path` to the project path. then normalize.
-function offset_svelte_path(svelte_path, input_path)
-  
-  -- offset from either input .qmd (if relative) or project dir (if absolute)
-  -- local offset_from = pandoc.system.get_working_directory()
-  local offset_from = "./"
-  if pandoc.path.is_relative(svelte_path) then
-    -- offset_from = pandoc.path.directory(input_path)
-    offset_from = pandoc.path.join({
-      offset_from,
-      pandoc.path.directory(input_path)
-    })
-  end
-  
-  -- join offset and svelt paths
-  local relative_path = pandoc.path.join({
-    offset_from,
-    pandoc.utils.stringify(svelte_path)
-  })
-  
-  -- normalize and return
-  local final_path = pandoc.path.normalize(relative_path)
-  return final_path
-end
+util = try_require({
+  "_extensions.jimjam-slam.sverto.util",
+  "_extensions.sverto.util",
+})
 
 input_paths = os.getenv("QUARTO_PROJECT_INPUT_FILES")
 
@@ -68,7 +32,7 @@ svelte_paths = {}
 -- for _, input_path in ipairs(input_paths) do
 for input_path in input_paths:gmatch("[^\n]+") do
 
-  local doc = pandoc.read(read_file(input_path), "markdown")
+  local doc = pandoc.read(util.read_file(input_path), "markdown")
 
   doc:walk {
     Meta = function(m)
@@ -78,20 +42,10 @@ for input_path in input_paths:gmatch("[^\n]+") do
         return nil
       end
 
-      local sverto_use
-      if pandoc.utils.type(m.sverto.use) == "List" then
-        sverto_use = m.sverto.use
-      elseif type(m.sverto.use) == "string" then
-        sverto_use = { m.sverto.use }
-      else
-        print(
-          "Sverto error: sverto.use key should be either a string path or " .. "a list of string paths, not " .. pandoc.utils.type(m.sverto.use))
-        return nil
-      end
-
+      local sverto_use = util.get_svelte_paths_from_meta(m)
       -- add each unique path, resolving relative project location
       for _, svelte_path in ipairs(sverto_use) do
-        local offset_path = offset_svelte_path(
+        local offset_path = util.offset_svelte_path(
           pandoc.utils.stringify(svelte_path),
           input_path)
         svelte_paths[offset_path] = offset_path
@@ -102,39 +56,12 @@ for input_path in input_paths:gmatch("[^\n]+") do
   
 end
 
--- load rollup.config.js from extension folder
-local rollup_config
-if file_exists("./_extensions/jimjam-slam/sverto/rollup.config.js") then
-  rollup_config = "./_extensions/jimjam-slam/sverto/rollup.config.js"
-elseif file_exists("./_extensions/sverto/rollup.config.js") then
-  rollup_config = "./_extensions/sverto/rollup.config.js"
-else
-  print("Error: Sverto extension files not found. " ..
-  "Is Sverto installed properly?")
-  os.exit(1)
-end
-
 -- now concat .svelte paths
 local svelte_path_string = ""
 for _, svelte_path in pairs(svelte_paths) do
   svelte_path_string = svelte_path_string ..  ":" .. svelte_path
 end
 
-cmd =
-  get_cmd_prefix() ..
-  "npm run build " ..
-  rollup_config .. " -- " ..
-  '--configQuartoOutPath="' .. os.getenv("QUARTO_PROJECT_OUTPUT_DIR") .. '" ' ..
-  '--configSvelteInPaths="' .. svelte_path_string .. '" ' ..
-  '--bundleConfigAsCjs'
-
-print("About to compile Svelte using cmd:")
-print(cmd)
-
-local svelteResult = os.execute(cmd)
-
-if svelteResult == nil or svelteResult == true then
-  print("Sverto pre-render finished!")
-else
-  print("Svelte compiler finished with code " .. tostring(svelteResult))
-end
+util.compile_svelte_files(
+  os.getenv("QUARTO_PROJECT_OUTPUT_DIR"),
+  svelte_path_string)
